+++ docs/about/whats_new.rst
@@ -16,6 +16,7 @@ v1.7dev
 - Bug fix (Index out of bounds causing ingestion failures)
 - Add "virtual products" (:pull:`522`, :pull:`597`) for multi-product loading
 - Support indexing data directly from HTTP/HTTPS/S3 URLs (:pull:`607`)
+- Stored `schema_version` added
 
 
 v1.6.1 (27 August 2018)